[Security] Move TraceableAuthenticator to security-http
